Blank sidebar on header selection, add phrase (#328)
* Changed to JsonConvert from JsonUtiltiy

JsonUtility should not be used.

* Blank sidebar on header selection, add phrase

Blanked the album sidebar info. It was easy to add number of sources but all other info seemed to be harder to get (afaik, meaning i'd have to play with the scanner which I didn't want to do). Also, added a the loadingPhrase text to the helpbar

* Revert "Blank sidebar on header selection, add phrase"

This reverts commit 4bcbc9210a3fae074447690b0c7d7ebae97778b3.

* Blank sidebar on header selection, add phrase

Blanked the album sidebar info. It was easy to add number of sources but all other info seemed to be harder to get (afaik, meaning i'd have to play with the scanner which I didn't want to do). Also, added a the loadingPhrase text to the helpbar

* Update SongSources.cs

* Special character replacement on generic lyrics.

For some reason, characters like the § were not parsed out for generic lyrics, only for vocal lyrics.