Two small changes (#352)
* Small typo update to Sidebar.cs

Sources was reporting i.genre instead of i.source

* Added the undertie character to Barlow

So you can rock out to Juanes with proper generic lyric support

* Fixed the warnings about font not being readable

If you go to credits or a few other places you will be warning spammed with "Unable to add the requested character to font asset atlas texture. Please make the texture readable." A total unity moment: 1) This can only be set when viewing the font atlas, while running, in the debug inspector, not the regular one. 2) They confuse reading and writing. 3) Why isn't this the default? Whatever, anyways, I set the Barlow Font to IsReadable true

* Revert "Fixed the warnings about font not being readable"

This reverts commit 7c9c7e5aa26e069d8380dd90618b93c4437ddcfb.

* Update Barlow.asset

* Missing script component removed

On the play scene background gameobject, there was a reference to an unknown missing script. Removed component

* Added GUI button to copy Json file

Added a button to get the Json currentsong file (like the text currentsong file)
In my attempt to add a new settings object (Labeled double button) I seemed to be adding a ton of code for a minor change. Instead I just added another button line.

* Two Small Changes

Fixed a missing word from the gui, added a toast for rescanning