player intro callouts (READ BEFORE MERGE)
Needs some love for sure. Couple things. For some reason it's really inconsistent. I try to only run it once when the song starts, and that only works sometimes? but when it works it works. Mic players don't have a callout tag. i have no idea how to add in for that. Also the FC animation I did is in here, just completely disabled, to avoid merge conflicts.