Added DMX Cue Change channel
Whenever the lighting cue changes, a DMX channel will be set to a value. Useful for programs that make light shows from incoming triggers.

This will be the last time the stagekit code will be used like this, next I'll redo how the cues are handled and uncouple that from the stagekit code directly so anything can listen to it better.