Improve volume scaling system
Apply exponential scaling onto only the final mix instead of every stem.

Note: we use exponential scaling for `<1` due to how human hearing works. The 50% that Bass provides in linear scaling does not actually give the impression of "50% volume" considering humans don't hear in linear space. It *has* to map to a slightly lower value for it to truly feel appropriate.

+ Change the cases for "processing threads" switch
+ Account for StemSettings changes for Master Volume
+ Update YARG.Core