ImageExtensions memory optimizations
Directly access the memory allocated by the Texture2D object. For Grayscale, that means using the texture's memory as the "result" variable. For DXT images, we do the byte swapping directly into that memory instead of in our own copy.
This also allows us to remove the branching for mipmap generation as we no longer need to creates a NativeArray struct around our copy of fixed memory.

+ Remove static Run functions for the grayscale jobs