Overhaul Texture Loading With New Image Functionality (#782)
* Utilize new image loading functionality for album art and source icons

+ requires extensions for YARGImage & RawImage that use unity structures and functions

* Dispose of YARGImage object when loading source icons

* Minor SongSource string optimizations

* Update to general sprite loading

+ Loads the source icon sprites of songs present in a user's library at scan time. This addresses the problem of source icon loading taking processing time when opening or scrolling through the music library menu.

+ Allows simplifications across the board to remove overreliance of async processing

* Update Background loading to match Core

+ Removes VenueInfo type

* Fix flipped background image

* Fix Difficultymenu source error

* Wrap AtomicSafetyHandle checks in ENABLE_UNITY_COLLECTIONS_CHECKS macros

Without these, any attempts to make a build would fail.

* Add STB2CSharp plugins

Thanks to Vinay Kapadia <dev@vkapadia.com> for the mac lib

* Remove TextureBuffer struct

* Silence `nullable` warnings

* Update Linux STB plugin

* ImageExtensions memory optimizations

Directly access the memory allocated by the Texture2D object. For Grayscale, that means using the texture's memory as the "result" variable. For DXT images, we do the byte swapping directly into that memory instead of in our own copy.
This also allows us to remove the branching for mipmap generation as we no longer need to creates a NativeArray struct around our copy of fixed memory.

+ Remove static Run functions for the grayscale jobs

* Add lock around image deletion & creation

It's possible that the threading could make two textures try to load within the same interval. Just in case, a single minor lock will keep *that* sequence sequential.

* Update MacOS plugin

* Call SongSources's LoadSprites from SongContainer.RunRefresh

Enforces the reset regardless of where a scan is called.

* Remove ViewObject Icon try-catch

* Remove ViewObject cancellation token behavior

* Revery downIcon caching

Unity caches the result under the hood already

* Remove incorrect comment

* Update YARG.Core