Changed callbacks to lamdas (#789)
* Linux Stage Kit hang on exit fix

For some reason, on linux the stage kit doesn't shut off properly. Hopefully this fixes it.

* Changed callbacks to lamdas

Removed unneeded  IsInitialized checks

* Update SacnInterpreter.cs

* Fixed start-up for DMX and Stage Kit

DMX and Stage Kit would work right in editor but not build. In build, on start, DMX and Stage Kit wouldn't get set to anything. This is fixed by removing the dependency on OnSceneLoaded and using the new MasterLightingInitalizer on every scene.

* Update StageKitInterpreter.cs

* Formatting changes

Also moved a function that was in the wrong spot

* Removed unneeded stuff

* PascalCase