Fix venues shaders on OSX (again)
Previously we were exporting shaders for OSX separately.
Unfortunately that does not seem to guarantee that all shader variants were exported.
To mitigate that we now export materials as well. However to avoid
doubling of resulting yarground size we replace all textures with dummy on export.

Loading side does not change