Fixes long standing bug with blending. (#972)
* Fixes long standing bug with blending.

Since I never had metadata for rb3 stuff, I never noticed the addition of blending (calling the same cue again). Since our auto-gen follows magma which does blending, auto-gen songs were having strange bugs in the stage kit because of this. Solution? just ignore a cue if same as the current one.

* Set post-processing back to default on scene change.

Otherwise it will stay at the last value from the song. 90% of the time this IS Default, but there are a few out there that don't behave.