Fix sustain rendering above frets.
Originally it was set to ignore depth as when highway curves it goes above sustain
and hides it. Sustain itself is not possible to curve without moving from
line renderer to some custom implementaion. But that led to sustain
drawing over the frets.

Instead of doing that, make sustain have standard depth test and instead
not write depth from the track shader so that track won't hide anything under it.