Added setting to choose how to display Discord rich presence (#1151)
* Added setting to turn on/off Discord rich presence

This setting has 3 options: Show, limited and hide. "Limited" will still show the activity card, but not the song you're currently playing.

* Fix for possible duplicate subscriptions to GameStateChange

Unsubscribes from GameStateChange in TryDispose() and resubscribes when calling Initialize()

* Fix for discord activity changing before localization was loaded

Initialize() is now only called once in LoadingScreen and sets _initialized to true in order to tell if it's safe to change the discord activity.

---------

Co-authored-by: Latif <56488217+latif-1005745@users.noreply.github.com>